+++ crates/evolrs_macros/src/crate_shape/method_traits/mod.rs
@@ -4,7 +4,7 @@ use syn::Ident;
 
 mod matmul;
 
-pub(crate) fn gen_methods(dims: usize, name: &Ident, dim_idents: &[Ident]) -> TokenStream {
+pub(crate) fn gen_methods(dims: i64, name: &Ident, dim_idents: &[Ident]) -> TokenStream {
     let matmul = matmul::matmul(dims, name, dim_idents);
     quote! {
         #matmul
